refactor(remote): extract transport-agnostic tool core (slice 0 step 2)
Every tool handler moves out of registerXTools(server, deps) closures
into src/core/ as ToolDefinition objects — (ctx, args) => result with
co-located zod schemas, arg types inferred via the SDK's own
ShapeOutput/ZodRawShapeCompat so signatures match registerTool exactly.
src/transports/stdio.ts is the thin adapter (single type-erasure site);
meta-tool and tool-groups stay adapter-side by design. ServerDeps
becomes SessionContext with a deprecated alias so callers compile
unchanged; evaluate handlers normalized to the full context. Test edits
are import paths only.

Kill-criterion count: 0 transport-dependent handlers beyond the
meta-tool. tools/list proven byte-identical to pre-refactor HEAD.
Suite: 987/987 green, golden fixtures untouched (I1 holds).

// ticktockbent
