+++ src/core/interaction-helpers.ts
