--- src/tools/session.ts
@@ -1,662 +1,0 @@
-import * as fs from "node:fs/promises";
-import * as path from "node:path";
-import { z } from "zod";
-import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
-import { logger } from "../utils/logger.js";
-import type { AutoSnapshotMode, DeviceType, DialogAutoDismiss } from "../types/config.js";
-import type { RegisteredTool } from "@modelcontextprotocol/sdk/server/mcp.js";
-import type { ToolDependencies } from "./tool-helpers.js";
-import {
-  ensureReady,
-  renderActivePage,
-  formatPageResponse,
-  handleToolError,
-  coercedBoolean,
-} from "./tool-helpers.js";
-
-const CookieSchema = z.object({
-  name: z.string().describe("Cookie name"),
-  value: z.string().describe("Cookie value"),
-  domain: z.string().describe("Cookie domain"),
-  path: z.string().optional().describe("Cookie path (default: '/')"),
-  secure: coercedBoolean.optional().describe("Secure flag"),
-  httpOnly: coercedBoolean.optional().describe("HttpOnly flag"),
-  sameSite: z.enum(["Strict", "Lax", "None"]).optional().describe("SameSite attribute"),
-});
-
-export function registerSessionTools(
-  server: McpServer,
-  deps: ToolDependencies,
-): Record<string, RegisteredTool> {
-  const tools: Record<string, RegisteredTool> = {};
-
-  // ─── charlotte_get_cookies ───
-  tools["charlotte_get_cookies"] = server.registerTool(
-    "charlotte_get_cookies",
-    {
-      description:
-        "Get cookies for the active page. Optionally filter by URL(s). Returns cookie name, value, domain, path, and flags.",
-      inputSchema: {
-        urls: z
-          .array(z.string())
-          .optional()
-          .describe(
-            "URLs to filter cookies by. If omitted, returns cookies for the current page URL.",
-          ),
-      },
-    },
-    async ({ urls }) => {
-      try {
-        await ensureReady(deps);
-        const page = deps.pageManager.getActivePage();
-
-        logger.info("Getting cookies", { urls });
-
-        const cookies = urls?.length ? await page.cookies(...urls) : await page.cookies();
-
-        return {
-          content: [
-            {
-              type: "text" as const,
-              text: JSON.stringify({
-                cookies: cookies.map((c) => ({
-                  name: c.name,
-                  value: c.value,
-                  domain: c.domain,
-                  path: c.path,
-                  expires: c.expires,
-                  httpOnly: c.httpOnly,
-                  secure: c.secure,
-                  sameSite: c.sameSite,
-                })),
-                count: cookies.length,
-              }),
-            },
-          ],
-        };
-      } catch (error: unknown) {
-        return handleToolError(error);
-      }
-    },
-  );
-
-  // ─── charlotte_clear_cookies ───
-  tools["charlotte_clear_cookies"] = server.registerTool(
-    "charlotte_clear_cookies",
-    {
-      description:
-        "Clear cookies from the browser. Optionally filter by name(s) to remove specific cookies. Without a filter, clears all cookies for the current page.",
-      inputSchema: {
-        names: z
-          .array(z.string())
-          .optional()
-          .describe("Cookie names to delete. If omitted, clears all cookies for the current page."),
-      },
-    },
-    async ({ names }) => {
-      try {
-        await ensureReady(deps);
-        const page = deps.pageManager.getActivePage();
-
-        logger.info("Clearing cookies", { names });
-
-        const currentCookies = await page.cookies();
-        const cookiesToDelete = names
-          ? currentCookies.filter((c) => names.includes(c.name))
-          : currentCookies;
-
-        await page.deleteCookie(...cookiesToDelete);
-
-        return {
-          content: [
-            {
-              type: "text" as const,
-              text: JSON.stringify({
-                success: true,
-                cleared: cookiesToDelete.length,
-                names: cookiesToDelete.map((c) => c.name),
-              }),
-            },
-          ],
-        };
-      } catch (error: unknown) {
-        return handleToolError(error);
-      }
-    },
-  );
-
-  // ─── charlotte_set_cookies ───
-  tools["charlotte_set_cookies"] = server.registerTool(
-    "charlotte_set_cookies",
-    {
-      description:
-        "Set cookies on the active page. Cookies persist for subsequent navigations within matching domains.",
-      inputSchema: {
-        cookies: z.array(CookieSchema).describe("Array of cookie objects to set"),
-      },
-    },
-    async ({ cookies }) => {
-      try {
-        await ensureReady(deps);
-        const page = deps.pageManager.getActivePage();
-
-        logger.info("Setting cookies", { count: cookies.length });
-
-        const puppeteerCookies = cookies.map((cookie) => ({
-          name: cookie.name,
-          value: cookie.value,
-          domain: cookie.domain,
-          path: cookie.path ?? "/",
-          secure: cookie.secure,
-          httpOnly: cookie.httpOnly,
-          sameSite: cookie.sameSite as "Strict" | "Lax" | "None" | undefined,
-        }));
-
-        await page.setCookie(...puppeteerCookies);
-
-        return {
-          content: [
-            {
-              type: "text" as const,
-              text: JSON.stringify({
-                success: true,
-                cookies_set: cookies.length,
-                details: cookies.map((c) => ({
-                  name: c.name,
-                  domain: c.domain,
-                  path: c.path ?? "/",
-                })),
-              }),
-            },
-          ],
-        };
-      } catch (error: unknown) {
-        return handleToolError(error);
-      }
-    },
-  );
-
-  // ─── charlotte_set_headers ───
-  tools["charlotte_set_headers"] = server.registerTool(
-    "charlotte_set_headers",
-    {
-      description:
-        "Set extra HTTP headers for subsequent requests. Headers persist for all navigations on the active page.",
-      inputSchema: {
-        headers: z
-          .record(z.string(), z.string())
-          .describe("Key-value header pairs (e.g. { 'Authorization': 'Bearer token' })"),
-      },
-    },
-    async ({ headers }) => {
-      try {
-        await ensureReady(deps);
-        const page = deps.pageManager.getActivePage();
-
-        logger.info("Setting extra HTTP headers", {
-          headerNames: Object.keys(headers),
-        });
-
-        await page.setExtraHTTPHeaders(headers);
-
-        return {
-          content: [
-            {
-              type: "text" as const,
-              text: JSON.stringify({
-                success: true,
-                headers_set: Object.keys(headers),
-              }),
-            },
-          ],
-        };
-      } catch (error: unknown) {
-        return handleToolError(error);
-      }
-    },
-  );
-
-  // ─── charlotte_configure ───
-  tools["charlotte_configure"] = server.registerTool(
-    "charlotte_configure",
-    {
-      description: "Configure Charlotte runtime settings. Changes take effect immediately.",
-      inputSchema: {
-        snapshot_depth: z
-          .number()
-          .optional()
-          .describe("Ring buffer size for snapshots (default: 50, min: 5, max: 500)"),
-        auto_snapshot: z
-          .enum(["every_action", "observe_only", "manual"])
-          .optional()
-          .describe(
-            '"every_action" (default) — snapshot after every tool, "observe_only" — only on observe, "manual" — only with explicit snapshot: true',
-          ),
-        screenshot_dir: z
-          .string()
-          .optional()
-          .describe(
-            "Directory for persistent screenshot artifacts. Changes take effect immediately; existing artifacts remain in the previous directory.",
-          ),
-        dialog_auto_dismiss: z
-          .enum(["none", "accept_alerts", "accept_all", "dismiss_all"])
-          .optional()
-          .describe(
-            'Auto-dismiss behavior for JS dialogs. "none" (default) queues for charlotte_dialog.',
-          ),
-        output_dir: z
-          .string()
-          .optional()
-          .describe(
-            "Directory for large tool output files (used by output_file parameter on observe, screenshot, console, requests). Created automatically if it doesn't exist.",
-          ),
-        include_iframes: coercedBoolean
-          .optional()
-          .describe(
-            "Include iframe content in page representations (default: false). Extracts accessibility tree, interactive elements, and content from child frames.",
-          ),
-        iframe_depth: z
-          .number()
-          .optional()
-          .describe("Maximum iframe nesting depth to traverse (default: 3, min: 1, max: 10)"),
-      },
-    },
-    async ({
-      snapshot_depth,
-      auto_snapshot,
-      screenshot_dir,
-      dialog_auto_dismiss,
-      output_dir,
-      include_iframes,
-      iframe_depth,
-    }) => {
-      try {
-        logger.info("Configuring Charlotte", { snapshot_depth, auto_snapshot, screenshot_dir });
-
-        if (snapshot_depth !== undefined) {
-          deps.snapshotStore.setDepth(snapshot_depth);
-          deps.config.snapshotDepth = Math.max(5, Math.min(500, snapshot_depth));
-        }
-
-        if (auto_snapshot !== undefined) {
-          deps.config.autoSnapshot = auto_snapshot as AutoSnapshotMode;
-        }
-
-        if (screenshot_dir !== undefined) {
-          // Validate against allowedWorkspaceRoot so that an agent cannot use charlotte_configure
-          // to silently repoint the screenshot directory outside the workspace. resolveOutputPath
-          // enforces containment at write time, but validating here catches the misconfiguration
-          // early and provides a clear error message. The agent is trusted to manage files within
-          // the workspace; this boundary is a bug-catcher, not a hard security sandbox.
-          const resolvedScreenshotDir = path.resolve(screenshot_dir);
-          const workspaceRoot =
-            deps.config.allowedWorkspaceRoot ??
-            (deps.config.outputDir ? path.resolve(deps.config.outputDir) : process.cwd());
-          if (
-            !resolvedScreenshotDir.startsWith(workspaceRoot + path.sep) &&
-            resolvedScreenshotDir !== workspaceRoot
-          ) {
-            throw new Error(
-              `screenshot_dir '${screenshot_dir}' resolves outside the allowed workspace root '${workspaceRoot}'. ` +
-                `Use a path within the workspace or update allowedWorkspaceRoot.`,
-            );
-          }
-          deps.config.screenshotDir = resolvedScreenshotDir;
-          await deps.artifactStore.setScreenshotDir(resolvedScreenshotDir);
-        }
-
-        if (dialog_auto_dismiss !== undefined) {
-          deps.config.dialogAutoDismiss = dialog_auto_dismiss as DialogAutoDismiss;
-        }
-
-        if (output_dir !== undefined) {
-          const resolvedOutputDir = path.resolve(output_dir);
-          // Validate against allowedWorkspaceRoot (same rationale as screenshot_dir above).
-          const workspaceRoot =
-            deps.config.allowedWorkspaceRoot ??
-            (deps.config.screenshotDir ? path.resolve(deps.config.screenshotDir) : process.cwd());
-          if (
-            !resolvedOutputDir.startsWith(workspaceRoot + path.sep) &&
-            resolvedOutputDir !== workspaceRoot
-          ) {
-            throw new Error(
-              `output_dir '${output_dir}' resolves outside the allowed workspace root '${workspaceRoot}'. ` +
-                `Use a path within the workspace or update allowedWorkspaceRoot.`,
-            );
-          }
-          deps.config.outputDir = resolvedOutputDir;
-          await fs.mkdir(resolvedOutputDir, { recursive: true });
-        }
-
-        if (include_iframes !== undefined) {
-          deps.config.includeIframes = include_iframes;
-        }
-
-        if (iframe_depth !== undefined) {
-          deps.config.iframeDepth = Math.max(1, Math.min(10, iframe_depth));
-        }
-
-        return {
-          content: [
-            {
-              type: "text" as const,
-              text: JSON.stringify({
-                success: true,
-                config: {
-                  snapshot_depth: deps.config.snapshotDepth,
-                  auto_snapshot: deps.config.autoSnapshot,
-                  screenshot_dir: deps.artifactStore.screenshotDir,
-                  dialog_auto_dismiss: deps.config.dialogAutoDismiss,
-                  output_dir: deps.config.outputDir ?? null,
-                  include_iframes: deps.config.includeIframes,
-                  iframe_depth: deps.config.iframeDepth,
-                },
-              }),
-            },
-          ],
-        };
-      } catch (error: unknown) {
-        return handleToolError(error);
-      }
-    },
-  );
-
-  // ─── charlotte_tabs ───
-  tools["charlotte_tabs"] = server.registerTool(
-    "charlotte_tabs",
-    {
-      description: "List all open browser tabs with their URLs, titles, and active status.",
-      inputSchema: {},
-    },
-    async () => {
-      try {
-        await ensureReady(deps);
-
-        const tabs = await deps.pageManager.listTabs();
-
-        return {
-          content: [
-            {
-              type: "text" as const,
-              text: JSON.stringify({ tabs }),
-            },
-          ],
-        };
-      } catch (error: unknown) {
-        return handleToolError(error);
-      }
-    },
-  );
-
-  // ─── charlotte_tab_open ───
-  tools["charlotte_tab_open"] = server.registerTool(
-    "charlotte_tab_open",
-    {
-      description:
-        "Open a new browser tab. Optionally navigate to a URL. The new tab becomes the active tab.",
-      inputSchema: {
-        url: z.string().optional().describe("URL to navigate to (default: blank page)"),
-      },
-    },
-    async ({ url }) => {
-      try {
-        await ensureReady(deps);
-
-        const tabId = await deps.pageManager.openTab(deps.browserManager, url);
-        logger.info("Opened new tab", { tabId, url });
-
-        const representation = await renderActivePage(deps, {
-          source: "action",
-        });
-
-        // Route through formatPageResponse so the page payload is stripped/size-
-        // capped like every other tool, with the new tab_id merged in (#188, #204).
-        return formatPageResponse(representation, {
-          extra: { tab_id: tabId },
-          maxResponseBytes: deps.config.limits.maxResponseBytes,
-        });
-      } catch (error: unknown) {
-        return handleToolError(error);
-      }
-    },
-  );
-
-  // ─── charlotte_tab_switch ───
-  tools["charlotte_tab_switch"] = server.registerTool(
-    "charlotte_tab_switch",
-    {
-      description:
-        "Switch to a different browser tab by its tab ID. Returns the page representation of the activated tab.",
-      inputSchema: {
-        tab_id: z.string().describe("ID of the tab to switch to"),
-      },
-    },
-    async ({ tab_id }) => {
-      try {
-        await ensureReady(deps);
-
-        await deps.pageManager.switchTab(tab_id);
-        logger.info("Switched to tab", { tab_id });
-
-        const representation = await renderActivePage(deps, {
-          source: "action",
-        });
-
-        return formatPageResponse(representation, {
-          maxResponseBytes: deps.config.limits.maxResponseBytes,
-        });
-      } catch (error: unknown) {
-        return handleToolError(error);
-      }
-    },
-  );
-
-  // ─── charlotte_tab_close ───
-  tools["charlotte_tab_close"] = server.registerTool(
-    "charlotte_tab_close",
-    {
-      description:
-        "Close a browser tab by its ID. If the closed tab was active, switches to the first remaining tab.",
-      inputSchema: {
-        tab_id: z.string().describe("ID of the tab to close"),
-      },
-    },
-    async ({ tab_id }) => {
-      try {
-        await ensureReady(deps);
-
-        await deps.pageManager.closeTab(tab_id);
-        logger.info("Closed tab", { tab_id });
-
-        const remainingTabs = await deps.pageManager.listTabs();
-
-        return {
-          content: [
-            {
-              type: "text" as const,
-              text: JSON.stringify({
-                success: true,
-                closed: tab_id,
-                remaining_tabs: remainingTabs,
-              }),
-            },
-          ],
-        };
-      } catch (error: unknown) {
-        return handleToolError(error);
-      }
-    },
-  );
-
-  // ─── charlotte_viewport ───
-
-  tools["charlotte_viewport"] = server.registerTool(
-    "charlotte_viewport",
-    {
-      description:
-        "Change the browser viewport dimensions. Use a device preset or specify custom width/height. Returns page representation at the new viewport size.",
-      inputSchema: {
-        width: z.number().min(1).optional().describe("Viewport width in pixels"),
-        height: z.number().min(1).optional().describe("Viewport height in pixels"),
-        device: z
-          .enum(["mobile", "tablet", "desktop"])
-          .optional()
-          .describe(
-            'Device preset (overrides width/height). "mobile" = 393×852, "tablet" = 768×1024, "desktop" = 1440×900',
-          ),
-      },
-    },
-    async ({ width, height, device }) => {
-      try {
-        await ensureReady(deps);
-        const page = deps.pageManager.getActivePage();
-        const { defaultViewport, deviceViewportPresets } = deps.config;
-
-        let viewportWidth: number;
-        let viewportHeight: number;
-
-        if (device) {
-          const preset = deviceViewportPresets[device as DeviceType];
-          viewportWidth = preset.width;
-          viewportHeight = preset.height;
-        } else if (width !== undefined && height !== undefined) {
-          viewportWidth = width;
-          viewportHeight = height;
-        } else {
-          viewportWidth = width ?? defaultViewport.width;
-          viewportHeight = height ?? defaultViewport.height;
-        }
-
-        logger.info("Setting viewport", {
-          width: viewportWidth,
-          height: viewportHeight,
-          device,
-        });
-
-        await page.setViewport({
-          width: viewportWidth,
-          height: viewportHeight,
-        });
-
-        const representation = await renderActivePage(deps, {
-          source: "action",
-        });
-
-        return formatPageResponse(representation, {
-          maxResponseBytes: deps.config.limits.maxResponseBytes,
-        });
-      } catch (error: unknown) {
-        return handleToolError(error);
-      }
-    },
-  );
-
-  // ─── charlotte_network ───
-
-  const THROTTLE_PRESETS: Record<
-    string,
-    {
-      offline: boolean;
-      downloadThroughput: number;
-      uploadThroughput: number;
-      latency: number;
-    }
-  > = {
-    "3g": {
-      offline: false,
-      downloadThroughput: (1.6 * 1024 * 1024) / 8,
-      uploadThroughput: (750 * 1024) / 8,
-      latency: 150,
-    },
-    "4g": {
-      offline: false,
-      downloadThroughput: (4 * 1024 * 1024) / 8,
-      uploadThroughput: (3 * 1024 * 1024) / 8,
-      latency: 20,
-    },
-    offline: {
-      offline: true,
-      downloadThroughput: 0,
-      uploadThroughput: 0,
-      latency: 0,
-    },
-    none: {
-      offline: false,
-      downloadThroughput: -1,
-      uploadThroughput: -1,
-      latency: 0,
-    },
-  };
-
-  tools["charlotte_network"] = server.registerTool(
-    "charlotte_network",
-    {
-      description:
-        "Configure network conditions for the active page. Set throttling presets, block URL patterns, or enable request logging.",
-      inputSchema: {
-        throttle: z
-          .enum(["3g", "4g", "offline", "none"])
-          .optional()
-          .describe(
-            'Network throttling preset. "3g" = slow, "4g" = fast mobile, "offline" = no network, "none" = disable throttling',
-          ),
-        block: z
-          .array(z.string())
-          .optional()
-          .describe(
-            'URL patterns to block (e.g. ["*.ads.com", "tracking.js"]). Pass empty array to clear.',
-          ),
-      },
-    },
-    async ({ throttle, block }) => {
-      try {
-        await ensureReady(deps);
-        const page = deps.pageManager.getActivePage();
-        // Use the cached long-lived CDP session rather than creating a fresh one: blocking
-        // rules die with their session, so the cached session (which lives as long as the
-        // page) is the correct lifetime — and a fresh session would be leaked (never
-        // detached). Network is NOT pre-enabled on this session (the render path doesn't
-        // need it), so enable it here explicitly; without it setBlockedURLs and
-        // emulateNetworkConditions are silent no-ops (#192). Idempotent per session.
-        const session = await deps.cdpSessionManager.getSession(page);
-        await session.send("Network.enable");
-
-        const appliedSettings: {
-          throttle?: string;
-          blocked_patterns?: string[];
-        } = {};
-
-        if (throttle !== undefined) {
-          const preset = THROTTLE_PRESETS[throttle];
-          await session.send("Network.emulateNetworkConditions", preset);
-          appliedSettings.throttle = throttle;
-          logger.info("Applied network throttling", { throttle });
-        }
-
-        if (block !== undefined) {
-          await session.send("Network.setBlockedURLs", { urls: block });
-          appliedSettings.blocked_patterns = block;
-          logger.info("Set blocked URL patterns", {
-            patternCount: block.length,
-          });
-        }
-
-        return {
-          content: [
-            {
-              type: "text" as const,
-              text: JSON.stringify({
-                success: true,
-                network: appliedSettings,
-              }),
-            },
-          ],
-        };
-      } catch (error: unknown) {
-        return handleToolError(error);
-      }
-    },
-  );
-
-  return tools;
-}
