+++ tests/integration/file-output.test.ts
@@ -10,14 +10,14 @@ import { ElementIdGenerator } from "../../src/renderer/element-id-generator.js";
 import { SnapshotStore } from "../../src/state/snapshot-store.js";
 import { ArtifactStore } from "../../src/state/artifact-store.js";
 import { createDefaultConfig } from "../../src/types/config.js";
-import type { ToolDependencies } from "../../src/tools/tool-helpers.js";
+import type { ToolDependencies } from "../../src/core/tool-helpers.js";
 import {
   renderActivePage,
   resolveOutputPath,
   writeOutputFile,
   writeBinaryOutputFile,
   stripEmptyFields,
-} from "../../src/tools/tool-helpers.js";
+} from "../../src/core/tool-helpers.js";
 
 const SIMPLE_FIXTURE = `file://${path.resolve(import.meta.dirname, "../fixtures/pages/simple.html")}`;
