+++ tests/integration/interaction.test.ts
@@ -11,9 +11,9 @@ import { ElementIdGenerator } from "../../src/renderer/element-id-generator.js";
 import { SnapshotStore } from "../../src/state/snapshot-store.js";
 import { ArtifactStore } from "../../src/state/artifact-store.js";
 import { createDefaultConfig } from "../../src/types/config.js";
-import type { ToolDependencies } from "../../src/tools/tool-helpers.js";
-import { renderActivePage, resolveElement } from "../../src/tools/tool-helpers.js";
-import { typeIntoElement } from "../../src/tools/interaction-helpers.js";
+import type { ToolDependencies } from "../../src/core/tool-helpers.js";
+import { renderActivePage, resolveElement } from "../../src/core/tool-helpers.js";
+import { typeIntoElement } from "../../src/core/interaction-helpers.js";
 import {
   setupMcpHarness,
   parseToolJson,
