+++ tests/integration/keyboard.test.ts
@@ -11,8 +11,8 @@ import { ElementIdGenerator } from "../../src/renderer/element-id-generator.js";
 import { SnapshotStore } from "../../src/state/snapshot-store.js";
 import { ArtifactStore } from "../../src/state/artifact-store.js";
 import { createDefaultConfig } from "../../src/types/config.js";
-import type { ToolDependencies } from "../../src/tools/tool-helpers.js";
-import { renderActivePage } from "../../src/tools/tool-helpers.js";
+import type { ToolDependencies } from "../../src/core/tool-helpers.js";
+import { renderActivePage } from "../../src/core/tool-helpers.js";
 import { pollUntil } from "../helpers/poll.js";
 
 const KEYBOARD_FIXTURE = `file://${path.resolve(import.meta.dirname, "../fixtures/pages/keyboard.html")}`;
