+++ tests/integration/observation.test.ts
@@ -10,13 +10,13 @@ import { ElementIdGenerator } from "../../src/renderer/element-id-generator.js";
 import { SnapshotStore } from "../../src/state/snapshot-store.js";
 import { ArtifactStore } from "../../src/state/artifact-store.js";
 import { createDefaultConfig } from "../../src/types/config.js";
-import type { ToolDependencies } from "../../src/tools/tool-helpers.js";
+import type { ToolDependencies } from "../../src/core/tool-helpers.js";
 import {
   renderActivePage,
   resolveElement,
   formatElementsResponse,
   waitForCompositorFrame,
-} from "../../src/tools/tool-helpers.js";
+} from "../../src/core/tool-helpers.js";
 import type { InteractiveElement, Bounds } from "../../src/types/page-representation.js";
 
 const SIMPLE_FIXTURE = `file://${path.resolve(import.meta.dirname, "../fixtures/pages/simple.html")}`;
