+++ tests/integration/popup.test.ts
@@ -12,9 +12,9 @@ import { ElementIdGenerator } from "../../src/renderer/element-id-generator.js";
 import { SnapshotStore } from "../../src/state/snapshot-store.js";
 import { ArtifactStore } from "../../src/state/artifact-store.js";
 import { createDefaultConfig } from "../../src/types/config.js";
-import type { ToolDependencies } from "../../src/tools/tool-helpers.js";
-import { renderActivePage } from "../../src/tools/tool-helpers.js";
-import { waitForPossibleNavigation } from "../../src/tools/interaction.js";
+import type { ToolDependencies } from "../../src/core/tool-helpers.js";
+import { renderActivePage } from "../../src/core/tool-helpers.js";
+import { waitForPossibleNavigation } from "../../src/core/interaction.js";
 import { pollUntil } from "../helpers/poll.js";
 
 const FIXTURES_DIR = path.resolve(import.meta.dirname, "../fixtures/pages");
