+++ tests/integration/wait-for.test.ts
@@ -25,9 +25,9 @@ import { ElementIdGenerator } from "../../src/renderer/element-id-generator.js";
 import { SnapshotStore } from "../../src/state/snapshot-store.js";
 import { ArtifactStore } from "../../src/state/artifact-store.js";
 import { createDefaultConfig } from "../../src/types/config.js";
-import type { ToolDependencies } from "../../src/tools/tool-helpers.js";
-import { renderActivePage } from "../../src/tools/tool-helpers.js";
-import { registerWaitForTools } from "../../src/tools/wait-for.js";
+import type { ToolDependencies } from "../../src/core/tool-helpers.js";
+import { renderActivePage } from "../../src/core/tool-helpers.js";
+import { registerWaitForTools } from "../../src/transports/stdio.js";
 
 const WAIT_FOR_FIXTURE = `file://${path.resolve(import.meta.dirname, "../fixtures/pages/wait-for.html")}`;
 const SIMPLE_FIXTURE = `file://${path.resolve(import.meta.dirname, "../fixtures/pages/simple.html")}`;
